Refinando o Model

+++ app/Core/Model.php
@@ -23,6 +23,10 @@ abstract class Model
 
     private $__storage = false;
 
+    protected $__protected_delete = false;
+
+    private $__protected_delete_column = 'exclusao_data';
+
     public function __construct($id = null){
         if(isset($id)){
             $this->load($id);
@@ -105,9 +109,14 @@ abstract class Model
 
     public function delete()
     {
+      
         if($this->__storage){
-            $sql = "DELETE FROM $this->table WHERE $this->pk = :$this->pk;";
-            $this->query($sql, [$this->pk => $this->{$this->pk}]);
+            if ($this->__protected_delete) {
+                $this->update([$this->__protected_delete_column => date('Y-m-d H:i:s')]);
+            }else{
+                $sql = "DELETE FROM $this->table WHERE $this->pk = :$this->pk;";
+                $this->query($sql, [$this->pk => $this->{$this->pk}]);
+            }
             $this->__storage = false;
             return true;
         }
@@ -152,12 +161,21 @@ abstract class Model
         $where = '';
         $data = [];
         if(count($this->where)>0){
-            $this->where[0][0] = 'WHERE';
+            $this->where[0][0] = '';
             foreach($this->where as $key => $w){
                 $where .= " $w[0] $w[1] $w[2] :w$key";
                 $data["w$key"] = $w[3];
             }
             $this->where = [];
+        }
+        if($this->__protected_delete){
+            if(empty($where)){
+                $where = " WHERE $this->__protected_delete_column IS NULL";
+            }else{
+                $where = " WHERE ($this->__protected_delete_column IS NULL) AND ($where)";
+            }
+        }else if(!empty($where)){
+            $where = " WHERE $where";
         }
         return [$where, $data];
     }
